# swagger.yml
openapi: 3.0.0

info:
  title: TrailService API
  version: 1.0.0
  description: >
    RESTful API for managing wellbeing trails.
    Authentication is delegated to an external COMP2001 Authenticator API.

servers:
  - url: /

tags:
  - name: Authentication
    description: User authentication
  - name: Trails
    description: Trail-related operations

paths:

  /login:
    post:
      tags: [Authentication]
      summary: User login
      operationId: login.login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        "200":
          description: Login successful
          
        "400":
          description: Invalid request format
       
        "401":
          description: Invalid email or password
       
        "500":
          description: Authentication service unavailable
       

  /trails:
    get:
      tags: [Trails]
      summary: Get all trails (basic info)
      operationId: trails.get_trails
      responses:
        "200":
          description: Successfully retrieved all trails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trail"
        "500":
          description: Server error
       

  /trails/detailed:
    get:
      tags: [Trails]
      summary: Get all trails with full details
      operationId: trails.get_trails_detailed
      responses:
        "200":
          description: Detailed trail list with related information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrailDetailed"
        "500":
          description: Server error
       

    post:
      tags: [Trails]
      summary: Create a new trail
      operationId: trails.create_trail
     
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrailCreate"
      responses:
        "201":
          description: Trail created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trail"
        "400":
          description: Invalid request body
        "500":
          description: Server error
       

  /trails/{trail_id}:
    parameters:
      - in: path
        name: trail_id
        required: true
        schema:
          type: string
        description: Trail ID (e.g., T000001)
        example: "T000001"

    get:
      tags: [Trails]
      summary: Get trail by ID with full details
      operationId: trails.get_trail
      responses:
        "200":
          description: Trail found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrailDetailed"
        "404":
          description: Trail not found
       
        "500":
          description: Server error
       

    put:
      tags: [Trails]
      summary: Update a trail
      operationId: trails.update_trail
     
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrailUpdate"
      responses:
        "200":
          description: Trail updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trail"
        "400":
          description: Invalid request body
       
        "404":
          description: Trail not found
       
        "500":
          description: Server error
       

    delete:
      tags: [Trails]
      summary: Delete a trail
      operationId: trails.delete_trail
      responses:
        "204":
          description: Trail deleted successfully
     
        "404":
          description: Trail not found
      
        "500":
          description: Server error

  /reference/locations:
  get:
    tags: [Reference Data]
    summary: Get all locations
    operationId: reference.get_locations
    responses:
      "200":
        description: List of available locations

/reference/routes:
  get:
    tags: [Reference Data]
    summary: Get all route types
    operationId: reference.get_routes

/reference/difficulties:
  get:
    tags: [Reference Data]
    summary: Get all difficulty levels
    operationId: reference.get_difficulties
        

components:
  schemas:
    Trail:
      type: object
      properties:
        trail_id:
          type: string
          example: "T000001"
        trail_name:
          type: string
          example: "Plymouth Waterfront Walk"
        description:
          type: string
          example: "Scenic coastal walk with harbor views"
        length:
          type: number
          format: double
          example: 5.2
        elevation_gain:
          type: integer
          example: 50
        estimated_time:
          type: number
          format: double
          example: 1.5
        visibility:
          type: string
          enum: [full, limited]
          example: "full"
        created_by:
          type: string
          example: "U000001"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        location_id:
          type: string
          example: "L000001"
        route_id:
          type: string
          example: "R000001"
        difficulty_id:
          type: string
          example: "D000002"
        points:
          type: array
          description: GPS waypoints for this trail
          items:
            $ref: "#/components/schemas/TrailPoint"

    TrailDetailed:
      allOf:
        - $ref: "#/components/schemas/Trail"
        - type: object
          properties:
            difficulty_level:
              type: string
              example: "Moderate"
            location_name:
              type: string
              example: "Plymouth"
            region:
              type: string
              example: "Devon"
            country:
              type: string
              example: "England"
            route_type:
              type: string
              example: "Loop"
            features:
              type: string
              example: "Coastal, Harbor, Historic"
            created_by_name:
              type: string
              example: "John Doe"
            creator_email:
              type: string
              example: "john.doe@example.com"

    TrailCreate:
      type: object
      required: [trail_name, visibility, difficulty_id, route_id, created_by]
      properties:
        trail_name:
          type: string
          minLength: 1
          maxLength: 100
          example: "New Coastal Trail"
        description:
          type: string
          example: "Beautiful seaside walk"
        length:
          type: number
          format: double
          minimum: 0.01
          maximum: 999.99
          example: 8.5
        elevation_gain:
          type: integer
          minimum: 0
          example: 150
        estimated_time:
          type: number
          format: double
          minimum: 0.01
          maximum: 99.99
          example: 2.5
        difficulty_id:
          type: string
          example: "D000002"
        route_id:
          type: string
          example: "R000001"
        location_id:
          type: string
          example: "L000001"
        visibility:
          type: string
          enum: [full, limited]
          example: "full"
        created_by:
          type: string
          example: "U000001"

    TrailUpdate:
      type: object
      properties:
        trail_name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Updated Trail Name"
        description:
          type: string
          example: "Updated description"
        visibility:
          type: string
          enum: [full, limited]
          example: "limited"
        length:
          type: number
          format: double
          minimum: 0.01
          maximum: 999.99
        elevation_gain:
          type: integer
          minimum: 0
        estimated_time:
          type: number
          format: double
          minimum: 0.01
          maximum: 99.99
        difficulty_id:
          type: string
        route_id:
          type: string
        location_id:
          type: string

    TrailPoint:
      type: object
      properties:
        trail_point_id:
          type: string
          example: "TP000001"
        trail_id:
          type: string
          example: "T000001"
        latitude:
          type: number
          format: double
          example: 50.351234
        longitude:
          type: number
          format: double
          example: -4.135678
        point_order:
          type: integer
          example: 1

   